+++ target_s3_csv/__init__.py
@@ -10,15 +10,16 @@ import shutil
 import sys
 from datetime import datetime
 import time
-
+import tempfile
 import singer
+
+from datetime import datetime
 from jsonschema import Draft7Validator, FormatChecker
 
 from target_s3_csv import s3
 from target_s3_csv import utils
 
-logger = singer.get_logger('target_s3_csv')
-
+logger = singer.get_logger()
 
 def emit_state(state):
     if state is not None:
@@ -29,127 +30,7 @@ def emit_state(state):
 
 
 # pylint: disable=too-many-locals,too-many-branches,too-many-statements
-# def persist_messages(messages, config):
-#     state = None
-#     schemas = {}
-#     key_properties = {}
-#     headers = {}
-#     validators = {}
-
-#     delimiter = config.get('delimiter', ',')
-#     quotechar = config.get('quotechar', '"')
-
-#     filenames = []
-#     now = datetime.now().strftime('%Y%m%dT%H%M%S')
-
-#     for message in messages:
-#         try:
-#             o = singer.parse_message(message).asdict()
-#         except json.decoder.JSONDecodeError:
-#             logger.error("Unable to parse:\n{}".format(message))
-#             raise
-#         message_type = o['type']
-#         if message_type == 'RECORD':
-#             if o['stream'] not in schemas:
-#                 raise Exception("A record for stream {}"
-#                                 "was encountered before a corresponding schema".format(o['stream']))
-
-#             # Validate record
-#             try:
-#                 validators[o['stream']].validate(utils.float_to_decimal(o['record']))
-#             except Exception as ex:
-#                 if type(ex).__name__ == "InvalidOperation":
-#                     logger.error("Data validation failed and cannot load to destination. RECORD: {}\n'multipleOf' validations that allows long precisions are not supported (i.e. with 15 digits or more). Try removing 'multipleOf' methods from JSON schema."
-#                     .format(o['record']))
-#                     raise ex
-
-#             record_to_load = o['record']
-#             if config.get('add_metadata_columns'):
-#                 record_to_load = utils.add_metadata_values_to_record(o, {})
-#             else:
-#                 record_to_load = utils.remove_metadata_values_from_record(o)
-
-#             filename = o['stream'] + '-' + now + '.csv'
-#             filename = os.path.expanduser(os.path.join(tempfile.gettempdir(), filename))
-#             if not filename in filenames:
-#                 filenames.append(filename)
-
-#             file_is_empty = (not os.path.isfile(filename)) or os.stat(filename).st_size == 0
-
-#             flattened_record = utils.flatten_record(record_to_load)
-
-#             if o['stream'] not in headers and not file_is_empty:
-#                 with open(filename, 'r') as csvfile:
-#                     reader = csv.reader(csvfile,
-#                                         delimiter=delimiter,
-#                                         quotechar=quotechar)
-#                     first_line = next(reader)
-#                     headers[o['stream']] = first_line if first_line else flattened_record.keys()
-#             else:
-#                 headers[o['stream']] = flattened_record.keys()
-
-#             with open(filename, 'a') as csvfile:
-#                 writer = csv.DictWriter(csvfile,
-#                                         headers[o['stream']],
-#                                         extrasaction='ignore',
-#                                         delimiter=delimiter,
-#                                         quotechar=quotechar)
-#                 if file_is_empty:
-#                     writer.writeheader()
-
-#                 writer.writerow(flattened_record)
-
-#             state = None
-#         elif message_type == 'STATE':
-#             logger.debug('Setting state to {}'.format(o['value']))
-#             state = o['value']
-#         elif message_type == 'SCHEMA':
-#             stream = o['stream']
-#             schemas[stream] = o['schema']
-#             if config.get('add_metadata_columns'):
-#                 schemas[stream] = utils.add_metadata_columns_to_schema(o)
-
-#             schema = utils.float_to_decimal(o['schema'])
-#             validators[stream] = Draft7Validator(schema, format_checker=FormatChecker())
-#             key_properties[stream] = o['key_properties']
-#         elif message_type == 'ACTIVATE_VERSION':
-#             logger.debug('ACTIVATE_VERSION message')
-#         else:
-#             logger.warning("Unknown message type {} in message {}"
-#                             .format(o['type'], o))
-
-#     # Upload created CSV files to S3
-#     for filename in filenames:
-#         compressed_file = None
-#         if config.get("compression") is None or config["compression"].lower() == "none":
-#             pass  # no compression
-#         else:
-#             if config["compression"] == "gzip":
-#                 compressed_file = f"{filename}.gz"
-#                 with open(filename, 'rb') as f_in:
-#                     with gzip.open(compressed_file, 'wb') as f_out:
-#                         logger.info(f"Compressing file as '{compressed_file}'")
-#                         shutil.copyfileobj(f_in, f_out)
-#             else:
-#                 raise NotImplementedError(
-#                     "Compression type '{}' is not supported. "
-#                     "Expected: 'none' or 'gzip'"
-#                     .format(config["compression"])
-#                 )
-#         s3.upload_file(compressed_file or filename,
-#                        config.get('s3_bucket'), config.get('s3_key_prefix', ''),
-#                        encryption_type=config.get('encryption_type'),
-#                        encryption_key=config.get('encryption_key'))
-
-#         # Remove the local file(s)
-#         os.remove(filename)
-#         if compressed_file:
-#             os.remove(compressed_file)
-
-#     return state
-
 def process_message_stream(input_message_stream, config):
-    state = None
     schemas = {}
     key_properties = {}
     headers = {}
@@ -173,17 +54,20 @@ def process_message_stream(input_message_stream, config):
             raise
         message_type = o['type']
         if message_type == 'RECORD':
-            if o['stream'] not in schemas:
+            stream_name = o['stream']
+
+            if stream_name not in schemas:
                 raise Exception("A record for stream {}"
-                                "was encountered before a corresponding schema".format(o['stream']))
+                                "was encountered before a corresponding schema".format(stream_name))
 
             # Validate record
             try:
-                validators[o['stream']].validate(utils.float_to_decimal(o['record']))
+                validators[stream_name].validate(utils.float_to_decimal(o['record']))
             except Exception as ex:
                 if type(ex).__name__ == "InvalidOperation":
-                    logger.error("Data validation failed and cannot load to destination. RECORD: {}\n'multipleOf' validations that allows long precisions are not supported (i.e. with 15 digits or more). Try removing 'multipleOf' methods from JSON schema."
-                    .format(o['record']))
+                    logger.error("Data validation failed and cannot load to destination. \n"
+                                 "'multipleOf' validations that allows long precisions are not supported"
+                                 " (i.e. with 15 digits or more). Try removing 'multipleOf' methods from JSON schema.")
                     raise ex
 
             record_to_load = o['record']
@@ -193,25 +77,29 @@ def process_message_stream(input_message_stream, config):
                 record_to_load = utils.remove_metadata_values_from_record(o)
 
             flattened_record = utils.flatten_record(record_to_load)
-            yield (o['stream'], flattened_record)
+            yield (o['stream'], flattened_record, message_type)
 
         elif message_type == 'STATE':
-            logger.info('Setting state to {}'.format(o['value']))
-            emit_state(o['value'])
+            logger.info(f'Received state from tap: {o["value"]}')
+            yield(None, o, message_type)
+            # logger.info('Setting state to {}'.format(o['value']))
+
         elif message_type == 'SCHEMA':
-            stream = o['stream']
-            schemas[stream] = o['schema']
+            stream_name = o['stream']
+            schemas[stream_name] = o['schema']
+
             if config.get('add_metadata_columns'):
-                schemas[stream] = utils.add_metadata_columns_to_schema(o)
+                schemas[stream_name] = utils.add_metadata_columns_to_schema(o)
 
             schema = utils.float_to_decimal(o['schema'])
-            validators[stream] = Draft7Validator(schema, format_checker=FormatChecker())
-            key_properties[stream] = o['key_properties']
+            validators[stream_name] = Draft7Validator(schema, format_checker=FormatChecker())
+            key_properties[stream_name] = o['key_properties']
         elif message_type == 'ACTIVATE_VERSION':
             logger.debug('ACTIVATE_VERSION message')
         else:
             logger.warning("Unknown message type {} in message {}".format(o['type'], o))
 
+    
 
 def main():
     now = datetime.now().strftime('%Y%m%dT%H%M%S')
@@ -229,22 +117,29 @@ def main():
     if len(config_errors) > 0:
         logger.error("Invalid configuration:\n   * {}".format('\n   * '.join(config_errors)))
         sys.exit(1)
-    s3.setup_aws_client(config)
+
+    # s3.setup_aws_client(config)
 
     uploaders = {}
     input_message_stream = io.TextIOWrapper(sys.stdin.buffer, encoding='utf-8')
     processed_message_stream = process_message_stream(input_message_stream, config)
-    for (stream_name, record) in processed_message_stream:
-        if stream_name not in uploaders:
-            s3_key = utils.get_target_key(stream_name, prefix=config.get('s3_key_prefix', ''), timestamp=now, naming_convention=config.get('naming_convention'))
-            uploaders[stream_name] = s3.S3MultipartUploader(config, stream_name, s3_key)
-        uploaders[stream_name].add_record(record)
-    for u in uploaders.values():
-        u.complete()
-
-    # input_messages = io.TextIOWrapper(sys.stdin.buffer, encoding='utf-8')
-    # state = persist_messages(input_messages, config)
-    # emit_state(state)
+    state_from_tap = None
+    try:
+        for (stream_name, record, message_type) in processed_message_stream:
+            if message_type == 'STATE':
+                state_from_tap = record
+            elif message_type == 'RECORD':
+                if stream_name not in uploaders:
+                    s3_key = utils.get_target_key(stream_name, prefix=config.get('s3_key_prefix', ''), timestamp=now, naming_convention=config.get('naming_convention'))
+                    uploaders[stream_name] = s3.S3MultipartUploader(config, stream_name, s3_key)
+                did_flush_records_to_target, record_count = uploaders[stream_name].add_record(record)
+                if did_flush_records_to_target and state_from_tap is not None:
+                    emit_state(state_from_tap)
+    finally:
+        for u in uploaders.values():
+            u.complete()
+        if state_from_tap is not None:
+            emit_state(state_from_tap)
 
     logger.debug("Exiting normally")