+++ target_s3_csv/utils.py
@@ -1,23 +1,21 @@
 #!/usr/bin/env python3
-
 import time
 import singer
 import json
 import re
-import collections
 import inflection
 
 from decimal import Decimal
 from datetime import datetime
+from collections.abc import MutableMapping
 
-logger = singer.get_logger('target_s3_csv')
+logger = singer.get_logger()
+
 
 def validate_config(config):
     """Validates config"""
     errors = []
     required_config_keys = [
-        'aws_access_key_id',
-        'aws_secret_access_key',
         's3_bucket'
     ]
 
@@ -58,6 +56,7 @@ def add_metadata_columns_to_schema(schema_message):
 
     return extended_schema_message
 
+
 def add_metadata_values_to_record(record_message, schema_message):
     """Populate metadata _sdc columns from incoming record message
     The location of the required attributes are fixed in the stream
@@ -72,6 +71,7 @@ def add_metadata_values_to_record(record_message, schema_message):
     extended_record['_sdc_table_version'] = record_message.get('version')
 
     return extended_record
+
 
 def remove_metadata_values_from_record(record_message):
     """Removes every metadata _sdc column from a given record message
@@ -103,14 +103,19 @@ def flatten_key(k, parent_key, sep):
 
     return sep.join(inflected_key)
 
-def flatten_record(d, parent_key=[], sep='__'):
+
+def flatten_record(d, parent_key=None, sep='__'):
     """
     """
+
+    if parent_key is None:
+        parent_key = []
+
     items = []
     for k in sorted(d.keys()):
         v = d[k]
         new_key = flatten_key(k, parent_key, sep)
-        if isinstance(v, collections.MutableMapping):
+        if isinstance(v, MutableMapping):
             items.extend(flatten_record(v, parent_key + [k], sep=sep).items())
         else:
             items.append((new_key, json.dumps(v) if type(v) is list else v))
@@ -124,12 +129,19 @@ def get_target_key(stream_name, prefix=None, timestamp=None, naming_convention=N
     if not timestamp:
         timestamp = datetime.now().strftime('%Y%m%dT%H%M%S')
     key = naming_convention
+    
+    # replace simple tokens
     for k, v in {
         '{stream}': stream_name,
         '{timestamp}': timestamp,
+        '{date}': datetime.now().strftime('%Y-%m-%d')
     }.items():
         if k in key:
             key = key.replace(k, v)
+
+    # replace dynamic tokens
+    # todo: replace dynamic tokens such as {date(<format>)} with the date formatted as requested in <format>
+
     if prefix:
         filename = key.split('/')[-1]
         key = key.replace(filename, f'{prefix}{filename}')