+++ CHANGELOG.md
@@ -1,3 +1,30 @@
+2.0.0 (2022-06-13)
+-------------------
+*Breaking changes*
+
+- Drop support for Python 3.6
+
+*Fixes*
+
+- FileNotFoundError when sync runs for more than a day 
+- State nullified after every record 
+
+1.5.0 (2021-08-13)
+-------------------
+- Add support for optional `aws_endpoint_url` to the configuration
+- Bumping dependencies
+
+1.4.0 (2020-09-09)
+-------------------
+
+- Add `temp_dir` option to write temporary files to custom location
+
+1.3.0 (2020-07-28)
+-------------------
+
+- Add `aws_session_token` and `aws_profile` options for Profile based authentication
+- Add `naming_convention` option with `{date}`, `{stream}` and `{timestamp}` tokens to create dynamic file names on S3
+
 1.2.1 (2020-06-17)
 -------------------